Day60/FastAPI/day-60-project/requirements.txt
fastapi
requests
pydantic
python-dotenv
jinja2
uvicorn
python-multipart